Change App component to a new Home component and make required changes in imports. Add Favorites page
