+++ src/components/Favorites.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Favorites() {
+  return <div>Favorite page</div>;
+}
